* Renamed PropertySet to PropertyRecord to avoid confusion with std::set
git-svn-id: svn://svn.berlios.de/dsa-hl@3941 4c79e8ff-cfd4-0310-af45-a38c79f83013

+++ rl/trunk/engine/startup/src/RastullahMain.cpp
@@ -174,7 +174,7 @@ INT WINAPI WinMain( HINSTANCE hInst, HINSTANCE, LPSTR strCmdLine, INT )
     int argc;
     LPWSTR* argList = CommandLineToArgvW(GetCommandLineW(), &argc);
     char** argv = new char*[argc];
-    for (size_t argIdx = 0; argIdx < argc; argIdx++)
+    for (int argIdx = 0; argIdx < argc; argIdx++)
     {
         size_t len = wcslen(argList[argIdx])+1;
         argv[argIdx] = new char[len];
